+++ src/main/org/apache/tools/ant/taskdefs/optional/depend/constantpool/FloatCPInfo.java
@@ -41,7 +41,7 @@ public class FloatCPInfo extends ConstantCPInfo {
      */
     @Override
     public void read(DataInputStream cpStream) throws IOException {
-        setValue(Float.valueOf(cpStream.readFloat()));
+        setValue(cpStream.readFloat());
     }
 
     /**
